refactor: timeline routes
