+++ internal/interfaces/handler.go
@@ -21,6 +21,5 @@ func (h *Handler) SetupRoutes() {
 		setUsersRoutesFrom(api)
 		setFollowsRoutesFrom(api)
 		setPostsRoutesFrom(api)
-		setActivityRoutesFrom(api)
 	}
 }
